Add sdc folder to server folder
