+++ db/sdc/seed-db-sdc-pg.js
@@ -1,8 +1,12 @@
+const { PerformanceObserver, performance } = require('perf_hooks');
 const videoData = require('../../helpers/sdc/dataSdcPg.js');// array of video data
 const db = require('../index.js');
 
+let initial = 0;
+let final = 0;
+
 let count = 0;
-const repeatTimes = 10; //this number * 10000 = total records inserted
+const repeatTimes = 999; //(this number * 10000) + 10000 = total records inserted
 
 const insertData = () => {
   return db('videos').insert(videoData)
@@ -16,6 +20,8 @@ const insertData = () => {
 
 db.schema.dropTableIfExists('videos')
   .then(() => {
+    initial = performance.now();
+    console.log('Seeding database...');
     return db.schema.createTable('videos', (t) => {
       t.increments('id').primary();
       t.string('video_url', 255);
@@ -26,7 +32,9 @@ db.schema.dropTableIfExists('videos')
     })
       .then(() => insertData())
       .then(() => {
-        console.log('database seeded');
+        console.log(`Database seeded, ${(repeatTimes * 10000) + 10000} records`);
+        final = performance.now();
+        console.log('Time elapsed: ', `${(final - initial) / 1000} seconds`)
         process.exit();
       })
       .catch((err) => {
