+++ package.json
@@ -16,9 +16,9 @@
   "scripts": {
     "react-dev": "webpack -d --watch",
     "build": "webpack -p",
-    "start": "node server/server.js",
+    "start": "node server/sdc/server.js",
     "seed": "node db/seed-db.js",
-    "seedSdcPg": "node db/sdc/seed-db-sdc-pg.js",
+    "seedSdcPg": "node --max-old-space-size=4096 db/sdc/seed-db-sdc-pg.js",
     "seedSdcMdb": "node --max-old-space-size=4096 db/sdc/seed-db-sdc-mdb.js",
     "test": "jest --detectOpenHandles"
   },
