+++ server/sdc/server.js
@@ -1,0 +1,6 @@
+const app = require('./index.js');
+
+const port = process.env.PORT || 3001;
+require('dotenv').config();
+
+app.listen(port, () => console.log(`listening on port ${port}`));
